requirements.txt
gym==0.26.2
gym-notices==0.0.8
hydra-core==1.3.2
omegaconf==2.3.0
opencv-python==4.9.0.80
pandas==2.0.3
pillow==10.3.0
seaborn==0.13.2
torch==2.2.2
torchvision==0.17.2
